Simplify (better: shorten) meta data creation
Use the function collect() directly in a call
to lapply() that loops over the attribute names
and convert the resulting list (of column
vectors) into a data frame.

Use setNames(nm = x) to create a vector in
which the elements are named with their
values. Using this in lapply() has the nice
side effect that the elements in the
returned list are named accordingly.