Allow to pass spatial object or no shape at all
You may now read a shape file on your own,
and pass the corresponding spatial object
(or e.g. a subset of it) to the funcion
read_monthly_data_over_shape(). You may also
set both, "file" and "shape" to NULL and should
receive data for the whole area of Germany.